🔨 [packages] Move type alias `ProviderStore` to `registry`
