+++ src/cutty/packages/adapters/storage.py
@@ -15,8 +15,8 @@ from typing import Optional
 from yarl import URL
 
 from cutty.packages.adapters.registry import defaultproviderfactories
-from cutty.packages.domain.providers import ProviderStore
 from cutty.packages.domain.registry import ProviderRegistry
+from cutty.packages.domain.registry import ProviderStore
 from cutty.packages.domain.stores import Store
